@@ -21,40 +21,40 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.2.10",
-    "@angular/cdk": "^15.2.9",
-    "@angular/common": "^15.2.10",
-    "@angular/compiler": "^15.2.10",
-    "@angular/core": "^15.2.10",
-    "@angular/forms": "^15.2.10",
-    "@angular/material": "^15.2.9",
-    "@angular/material-date-fns-adapter": "^15.2.9",
-    "@angular/platform-browser": "^15.2.10",
-    "@angular/platform-browser-dynamic": "^15.2.10",
-    "@angular/router": "^15.2.10",
+    "@angular/animations": "^16.2.12",
+    "@angular/cdk": "^16.2.14",
+    "@angular/common": "^16.2.12",
+    "@angular/compiler": "^16.2.12",
+    "@angular/core": "^16.2.12",
+    "@angular/forms": "^16.2.12",
+    "@angular/material": "^16.2.14",
+    "@angular/material-date-fns-adapter": "^16.2.14",
+    "@angular/platform-browser": "^16.2.12",
+    "@angular/platform-browser-dynamic": "^16.2.12",
+    "@angular/router": "^16.2.12",
     "@azure/msal-browser": "3.2.0",
-    "@enigmatry/entry-components": "15.1.0-preview",
-    "@enigmatry/entry-form": "15.1.0-preview",
-    "@microsoft/applicationinsights-web": "^3.0.7",
-    "@ngx-formly/core": "6.1.7",
-    "@ngx-formly/material": "6.1.7",
+    "@enigmatry/entry-components": "16.0.0",
+    "@enigmatry/entry-form": "16.0.0",
+    "@microsoft/applicationinsights-web": "3.1.0",
+    "@ngx-formly/core": "6.3.0",
+    "@ngx-formly/material": "6.3.0",
     "date-fns": "^2.30.0",
     "rxjs": "^7.8.1",
     "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.13.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.10",
-    "@angular/cli": "^15.2.10",
-    "@angular/compiler-cli": "^15.2.10",
-    "@angular/localize": "^15.2.10",
-    "@enigmatry/eslint-config": "15.0.0",
-    "@enigmatry/scss-foundation": "15.0.0",
-    "@enigmatry/stylelint-config": "15.0.0",
-    "@types/node": "^16.4.3",
-    "@types/jest": "28.1.3",
-    "jest": "28.1.3",
-    "jest-preset-angular": "^12.2.6",
+    "@angular-devkit/build-angular": "^16.2.12",
+    "@angular/cli": "^16.2.12",
+    "@angular/compiler-cli": "^16.2.12",
+    "@angular/localize": "^16.2.12",
+    "@enigmatry/eslint-config": "16.0.0",
+    "@enigmatry/scss-foundation": "16.0.0",
+    "@enigmatry/stylelint-config": "16.0.0",
+    "@types/node": "^16.18.86",
+    "@types/jest": "29.5.12",
+    "jest": "29.7.0",
+    "jest-preset-angular": "14.0.3",
     "json-sort-cli": "^4.0.0",
     "npm-audit-reporter-teamcity": "0.4.0",
     "nswag": "^14.0.0",
